Some refactoring and implementing image analysis service

+++ WebTranslator/Models/TranslationDTOs/Vision/BoundingBox.cs
@@ -2,9 +2,9 @@
 {
     public class BoundingBox
     {
-        public int X { get; set; }
-        public int Y { get; set; }
-        public int Width { get; set; }
-        public int Height { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
+        public double Width { get; set; }
+        public double Height { get; set; }
     }
 }
